[Silver IV] Title: 스케이트 연습, Time: 64 ms, Memory: 5932 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/28324. 스케이트 연습/스케이트 연습.cc
@@ -14,7 +14,7 @@ int main() {
 		std::cin >> v[i];
 	}
 	
-	int sum = 1;
+	long long sum = 1;
 	for (int i = n - 2; i >= 0; i--) {
 		if (v[i] < ans[i + 1] + 1) {
 			ans[i] = v[i];
